cleaned up warnings

+++ src/core/events.c
@@ -20,9 +20,9 @@ void event_system_initialize()
 void event_system_destroy()
 {
     INFO("Destroying event system...");
-    u64 events_size = array_get_length(state.events);
+    s32 events_size = (s32)array_get_length(state.events);
 
-    for (u32 i = 0; i < events_size; i++)
+    for (s32 i = 0; i < events_size; i++)
     {
         if (state.events[i].listeners)
         {
@@ -42,19 +42,19 @@ void event_register(event_type type, fp_on_event fp_on_event)
         return;
     }
 
-    u64    events_size = array_get_length(state.events);
+    s32    events_size = (s32)array_get_length(state.events);
     event *found_event = 0;
 
     // INFO: check if we have already registered for this event; I think this is ugly and I should make it better but atm idk how
-    for (u32 i = 0; i < events_size; i++)
+    for (s32 i = 0; i < events_size; i++)
     {
         if (state.events[i].type == type)
         {
             found_event = &state.events[i];
             if (state.events[i].listeners != 0)
             {
-                u64 listeners_size = array_get_length(state.events[i].listeners);
-                for (u32 j = 0; j < listeners_size; j++)
+                s64 listeners_size = (s32)array_get_length(state.events[i].listeners);
+                for (s32 j = 0; j < listeners_size; j++)
                 {
                     if (state.events[i].listeners[j].callback == fp_on_event)
                     {
@@ -92,7 +92,7 @@ void event_fire(event_type type, event_context context)
         return;
     }
 
-    u64 events_size = array_get_length(state.events);
+    s32 events_size = (s32)array_get_length(state.events);
 
     switch (type)
     {
